Add exception handling for starting/stopping a timer (#9)
* Add new method 'unfinished_entry_present()'

Checks if a timer has been started but not yet stopped.

* Add UnfinishedEntryPresentException class

* Throw exception when starting/stopping timer was invalid

Trying to start a new timer when one already exists will throw an exception. Trying to stop a timer when noone was started will throw an exception

* Add test checking for thrown exception with timer starting

* Implement exception hadnling in start() function

* Add test checking for thrown exception with timer stopping

* Implement exception handling in stop() function

* Refactor name of exception to 'InvalidTimerModification'

Old name 'UnfinishedEntryPresentException' was not good because the finish_created_entry()-method would raise if there is no unfinished entry present. Name was misleading