Fix two bugs when setting up test values
First: Test values were starting one/two days before today; should start today.
Second: Internally, the algorithm was generating twice as much dates as needed, because it generated the requested number of dates and doubled it later to create two entries per date. Fixed it because it was wasted memory.